clock: adjtimex/clock_adjtime mit persistenter NTP-State, CPUTIME-Clocks, EPERM
Linux-ABI-konforme adjtimex(2)/clock_adjtime(2):
- ntp_state haelt offset/freq/maxerror/esterror/status/constant/tick/tai
- SET-Modes schreiben persistent, READ-only-Calls (modes=0 / ADJ_OFFSET_SS_READ)
  ignorieren capability, jeder andere Mode-Bit erfordert euid=0 (-> EPERM)
- Unbekannte Mode-Bits -> EINVAL (ADJ_ADJTIME nur als SINGLESHOT/SS_READ)
- ADJ_TICK: [9000,11000]us, ADJ_FREQUENCY: +-32.768e6
- time.tv_sec/tv_usec werden bei jedem Call aus rtc_epoch_sec+timer_ms gefuellt
- Rueckgabewert: TIME_OK / TIME_INS / TIME_DEL / TIME_ERROR (STA_UNSYNC default)

clock_gettime(CLOCK_PROCESS_CPUTIME_ID / CLOCK_THREAD_CPUTIME_ID):
- Aus process->cpu_time_ticks (1000Hz-Akku), als ns ausgeliefert
- clock_getres akzeptiert diese Clock-IDs mit 1ms-Aufloesung

clock_nanosleep(CLOCK_THREAD_CPUTIME_ID) -> -ENOTSUP (Linux-konform)

clock_settime:
- euid!=0 -> -EPERM, sonst rtc_epoch_sec = kts.tv_sec - uptime_sec
- rtc_epoch_sec wird int64_t (vorher uint64_t, Rueckwaerts-Set haette
  unterlaufen). Alle Consumer (timer.c, sys_time.c, ipc/sysv_ipc.c,
  sys/dispatch.c, sys/stubs.c) auf signed umgestellt.

procfs /proc/self/stat utime/stime:
- Vorher immer 0 -> clock_gettime01-Setup-Loop `while (utime==0)` haengte.
- utime/stime in USER_HZ-Jiffies aus cpu_time_ticks (80/20 user/system Split).

Alle ADJ_*/STA_*/TIME_*-Konstanten in include/linux/time.h zentralisiert.

ktest 2694/0 (+32 neue Tests fuer adjtimex-Roundtrip, CPUTIME-Clocks,
EFAULT-Matrix, clock_settime recede, clock_nanosleep-ENOTSUP).

LTP: adjtimex02, clock_adjtime01/02, clock_gettime01/02 jetzt PASS (5 Tests
die vorher FAIL waren). Verbleibende clock_*-Fails: gettime03/nanosleep03
brauchen CLONE_NEWTIME (time namespaces), gettime04 tight-loop-Latenz,
nanosleep01/02 BAD_TS_ADDR_REM EFAULT-Fehlerpfad.

+++ src/arch/x86_64/timer/timer.c
@@ -181,7 +181,7 @@ uint64_t timer_ms(void) {
 }
 
 uint32_t timer_epoch_sec(void) {
-    return (uint32_t)(rtc_epoch_sec + timer_ms() / 1000);
+    return (uint32_t)(rtc_epoch_sec + (int64_t)(timer_ms() / 1000));
 }
 
 /* Kernel-only non-preemptible delay. For HW init timing (SMP SIPI, device
@@ -200,7 +200,7 @@ void timer_sleep_ms(uint32_t ms) {
 
 /* ── CMOS RTC → Unix epoch ────────────────────── */
 
-uint64_t rtc_epoch_sec = 0;
+int64_t rtc_epoch_sec = 0;
 
 static uint8_t cmos_read(uint8_t reg) {
     arch_outb(0x70, reg);
@@ -248,12 +248,12 @@ void rtc_init(void) {
 
     int full_year = 2000 + (int)year;
     uint64_t days = days_since_epoch(full_year, (int)mon, (int)day);
-    rtc_epoch_sec = days * 86400 + (uint64_t)hour * 3600
-                  + (uint64_t)min * 60 + (uint64_t)sec;
+    rtc_epoch_sec = (int64_t)(days * 86400 + (uint64_t)hour * 3600
+                  + (uint64_t)min * 60 + (uint64_t)sec);
 
     serial_puts("RTC: ");
     char tmp[20]; int ti = 0;
-    uint64_t v = rtc_epoch_sec;
+    uint64_t v = (uint64_t)rtc_epoch_sec;
     do { tmp[ti++] = '0' + v % 10; v /= 10; } while (v);
     while (ti--) serial_putchar(tmp[ti]);
     serial_puts(" epoch sec\n");
